ci(p1): blocking fmt+clippy+test gates (Linux); fix all clippy errors
P1.2: Add lint.yml — cargo fmt --check + cargo clippy -D warnings as
required blocking checks on ubuntu-latest. autofix.yml remains advisory
(auto-committing nightly fixer) but is no longer the source of truth for
gate status.

P1.3: Add test.yml — cargo nextest run --all-features --workspace as a
required blocking check on ubuntu-latest via taiki-e/install-action.
Both jobs use Swatinem/rust-cache and Linux only (no macOS/Windows
billed runners).

Fix all pre-existing clippy errors across 9 crates so the gate starts
green: doc-comment formatting (forge_similarity, forge_drift, forge3d),
missing Default impls (HashOnlyProvider, LocalFastembedProvider,
DriftIndex), derivable impls (FontWeight, BellStyle, GlobalSettings),
unused import / variable / dead-code (forge_pheno_winterminal,
forge3d, forge_app), disallowed String::from_utf8_lossy →
bstr::ByteSlice::to_str_lossy (forge_mux), nested-if collapse and
useless format! (forge_pheno_shell), needless borrow (forge_main),
unneeded return / redundant cast (forge3d).

+++ crates/forge3d/src/protocol.rs
@@ -220,7 +220,7 @@ mod tests {
     #[tokio::test]
     async fn decode_frame_rejects_oversize_header() {
         // Build a frame whose header advertises > MAX_FRAME_BYTES.
-        let header = (MAX_FRAME_BYTES as u32 + 1).to_be_bytes();
+        let header = (MAX_FRAME_BYTES + 1).to_be_bytes();
         let bytes: Vec<u8> = header.iter().chain(&[0u8; 0]).copied().collect();
         let mut cursor: &[u8] = &bytes;
         let err = decode_frame(&mut cursor).await.unwrap_err();
